Update add request

+++ src/main/java/com/example/Travel/io/repositories/ImageRepository.java
@@ -5,5 +5,5 @@ import com.example.Travel.io.Model.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ImageRepository extends JpaRepository<Image,Long> {
-    Image findByClientIdClient(Integer ClientIdClient);
+    Image findByClientIdClient(Long ClientIdClient);
 }
